Reorganize FileAdapter methods in traits
