+++ src/Phug/Renderer/Renderer/Partial/RenderingFileTrait.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace Phug\Renderer\Partial;
+
+trait RenderingFileTrait
+{
+    /**
+     * File currently rendering.
+     *
+     * @var string
+     */
+    private $renderingFile;
+
+    /**
+     * Get file currently rendering.
+     *
+     * @return string
+     */
+    public function getRenderingFile()
+    {
+        return $this->renderingFile;
+    }
+}
